+++ SwiftUI_Pull_to_Refresh/Views/UI.swift
@@ -25,8 +25,10 @@ struct UI {
 	
 	struct Image {
 		
-		static let background = SwiftUI.Image("WorldMap")
 		static let blur = CGFloat(4)
+		static let background = SwiftUI.Image("WorldMap")
+		static let opaqueBackground = SwiftUI.Image("Opaque World Map")
+		static let opaqueBackgroundWithBlur = SwiftUI.Image("Opaque World Map with Blur")
 	}
 	
 	struct Color {